myproject-main(1)/myproject-main/requirements.txt
Flask
gunicorn
requests
google.oauth2
google.analytics.data_v1beta